feat: optional app export as spreadsheet (#4322)
* feat: optional app export as spreadsheet

* fix: update per emily

* fix: updates per morgan
