+++ api/src/modules/application.module.ts
@@ -1,17 +1,18 @@
 import { Module } from '@nestjs/common';
 import { ApplicationController } from '../controllers/application.controller';
+import { ApplicationExporterModule } from './application-exporter.module';
 import { ApplicationService } from '../services/application.service';
-import { GeocodingService } from '../services/geocoding.service';
-import { PrismaModule } from './prisma.module';
-import { PermissionModule } from './permission.module';
 import { EmailModule } from './email.module';
+import { GeocodingService } from '../services/geocoding.service';
 import { ListingModule } from './listing.module';
 import { MultiselectQuestionModule } from './multiselect-question.module';
-import { ApplicationCsvExporterService } from '../services/application-csv-export.service';
+import { PermissionModule } from './permission.module';
+import { PrismaModule } from './prisma.module';
 import { UnitTypeModule } from './unit-type.module';
 
 @Module({
   imports: [
+    ApplicationExporterModule,
     PrismaModule,
     EmailModule,
     ListingModule,
@@ -20,11 +21,7 @@ import { UnitTypeModule } from './unit-type.module';
     UnitTypeModule,
   ],
   controllers: [ApplicationController],
-  providers: [
-    ApplicationService,
-    GeocodingService,
-    ApplicationCsvExporterService,
-  ],
+  providers: [ApplicationService, GeocodingService],
   exports: [ApplicationService],
 })
 export class ApplicationModule {}
