+++ api/src/modules/lottery.module.ts
@@ -1,16 +1,18 @@
 import { Logger, Module } from '@nestjs/common';
 import { SchedulerRegistry } from '@nestjs/schedule';
+import { ApplicationExporterModule } from './application-exporter.module';
 import { ConfigService } from '@nestjs/config';
+import { EmailModule } from './email.module';
+import { ListingModule } from './listing.module';
 import { LotteryController } from '../controllers/lottery.controller';
 import { LotteryService } from '../services/lottery.service';
-import { PrismaModule } from './prisma.module';
-import { PermissionModule } from './permission.module';
-import { ListingModule } from './listing.module';
 import { MultiselectQuestionModule } from './multiselect-question.module';
-import { EmailModule } from './email.module';
+import { PermissionModule } from './permission.module';
+import { PrismaModule } from './prisma.module';
 
 @Module({
   imports: [
+    ApplicationExporterModule,
     PrismaModule,
     ListingModule,
     EmailModule,
