+++ api/test/unit/services/application-exporter.service.spec.ts
@@ -5,9 +5,10 @@ import { MultiselectQuestionsApplicationSectionEnum } from '@prisma/client';
 import { HttpModule } from '@nestjs/axios';
 import { Request as ExpressRequest, Response } from 'express';
 import { PrismaService } from '../../../src/services/prisma.service';
-import { ApplicationCsvExporterService } from '../../../src/services/application-csv-export.service';
-import { MultiselectQuestionService } from '../../../src/services/multiselect-question.service';
+import { ApplicationCsvQueryParams } from '../../../src/dtos/applications/application-csv-query-params.dto';
 import { User } from '../../../src/dtos/users/user.dto';
+import { ApplicationExporterService } from '../../../src/services/application-exporter.service';
+import { MultiselectQuestionService } from '../../../src/services/multiselect-question.service';
 import { mockApplicationSet } from './application.service.spec';
 import { mockMultiselectQuestion } from './multiselect-question.service.spec';
 import { ListingService } from '../../../src/services/listing.service';
@@ -21,15 +22,15 @@ import { SchedulerRegistry } from '@nestjs/schedule';
 import { GoogleTranslateService } from '../../../src/services/google-translate.service';
 import { unitTypeToReadable } from '../../../src/utilities/application-export-helpers';
 
-describe('Testing application CSV export service', () => {
-  let service: ApplicationCsvExporterService;
+describe('Testing application export service', () => {
+  let service: ApplicationExporterService;
   let prisma: PrismaService;
   let permissionService: PermissionService;
 
   beforeAll(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
-        ApplicationCsvExporterService,
+        ApplicationExporterService,
         PrismaService,
         MultiselectQuestionService,
         ListingService,
@@ -52,8 +53,8 @@ describe('Testing application CSV export service', () => {
       imports: [HttpModule],
     }).compile();
 
-    service = module.get<ApplicationCsvExporterService>(
-      ApplicationCsvExporterService,
+    service = module.get<ApplicationExporterService>(
+      ApplicationExporterService,
     );
     prisma = module.get<PrismaService>(PrismaService);
     permissionService = module.get<PermissionService>(PermissionService);
@@ -94,13 +95,13 @@ describe('Testing application CSV export service', () => {
       },
     ]);
 
-    const exportResponse = await service.exportFile(
+    const exportResponse = await service.csvExport(
       { user: requestingUser } as unknown as ExpressRequest,
       {} as unknown as Response,
       {
         id: randomUUID(),
         includeDemographics: false,
-      },
+      } as unknown as ApplicationCsvQueryParams,
     );
 
     const headerRow =
@@ -153,10 +154,13 @@ describe('Testing application CSV export service', () => {
         ),
       ]);
 
-    const exportResponse = await service.exportFile(
+    const exportResponse = await service.csvExport(
       { user: requestingUser } as unknown as ExpressRequest,
       {} as unknown as Response,
-      { id: 'test', includeDemographics: true },
+      {
+        id: 'test',
+        includeDemographics: true,
+      } as unknown as ApplicationCsvQueryParams,
     );
 
     const headerRow =
@@ -221,10 +225,10 @@ describe('Testing application CSV export service', () => {
     jest
       .spyOn({ unitTypeToReadable }, 'unitTypeToReadable')
       .mockReturnValue('Studio');
-    const exportResponse = await service.exportFile(
+    const exportResponse = await service.csvExport(
       { user: requestingUser } as unknown as ExpressRequest,
       {} as unknown as Response,
-      { id: randomUUID() },
+      { id: randomUUID() } as unknown as ApplicationCsvQueryParams,
     );
 
     const mockedStream = new PassThrough();
@@ -284,13 +288,13 @@ describe('Testing application CSV export service', () => {
     jest
       .spyOn({ unitTypeToReadable }, 'unitTypeToReadable')
       .mockReturnValue('Studio');
-    const exportResponse = await service.exportFile(
+    const exportResponse = await service.csvExport(
       { user: requestingUser } as unknown as ExpressRequest,
       {} as unknown as Response,
       {
         id: randomUUID(),
         timeZone: 'America/New_York',
-      },
+      } as unknown as ApplicationCsvQueryParams,
     );
 
     const mockedStream = new PassThrough();
