+++ api/test/unit/services/lottery.service.spec.ts
@@ -11,7 +11,7 @@ import { HttpModule } from '@nestjs/axios';
 import Excel from 'exceljs';
 import { Request as ExpressRequest, Response } from 'express';
 import { PrismaService } from '../../../src/services/prisma.service';
-import { ApplicationCsvExporterService } from '../../../src/services/application-csv-export.service';
+import { ApplicationExporterService } from '../../../src/services/application-exporter.service';
 import { MultiselectQuestionService } from '../../../src/services/multiselect-question.service';
 import { User } from '../../../src/dtos/users/user.dto';
 import {
@@ -57,7 +57,7 @@ describe('Testing lottery service', () => {
   beforeAll(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
-        ApplicationCsvExporterService,
+        ApplicationExporterService,
         PrismaService,
         MultiselectQuestionService,
         ListingService,
