+++ sites/partners/.env.template
@@ -23,3 +23,5 @@ API_PASS_KEY="some-key-here"
 
 # feature toggles
 SHOW_LOTTERY=FALSE
+# controls whethere the application export is a spreadsheet (TRUE) or a csv (FALSE)
+APPLICATION_EXPORT_AS_SPREADSHEET=FALSE
