+++ sites/partners/src/pages/listings/[id]/lottery.tsx
@@ -30,7 +30,7 @@ import { logger } from "../../../logger"
 import {
   useFlaggedApplicationsMeta,
   useLotteryActivityLog,
-  useLotteryExport,
+  useSpreadsheetExport,
 } from "../../../lib/hooks"
 dayjs.extend(advancedFormat)
 
@@ -64,12 +64,13 @@ const Lottery = (props: { listing: Listing | undefined }) => {
   const includeDemographicsPartner =
     profile?.userRoles?.isPartner && listingJurisdiction?.enablePartnerDemographics
 
-  const { onExport, exportLoading } = useLotteryExport(
+  const { onExport, exportLoading } = useSpreadsheetExport(
     listing?.id,
     (profile?.userRoles?.isAdmin ||
       profile?.userRoles?.isJurisdictionalAdmin ||
       includeDemographicsPartner) ??
-      false
+      false,
+    true
   )
   const { data } = useFlaggedApplicationsMeta(listing?.id)
   const { lotteryActivityLogData } = useLotteryActivityLog(listing?.id)
